correcting the authentication system

+++ src/app/auth/auth.service.ts
@@ -50,49 +50,59 @@ export class AuthService {
     this.isAuthenticated$ = this.user$.pipe(map((res) => Boolean(this.CookieService.get('auth') && res.id)));
     if (this.CookieService.get('auth')) {
       this.getUser()
+    } else {
+      this.logout();
+      router.navigate(['/auth/login']);
     }
   }
 
   signIn(data: SignIn) {
     return this.api.post<{token:String}>(`/user/signin`, data).pipe(
       tap((res: any) => {
-        this.CookieService.deleteAll("auth")
-        this.CookieService.set('auth', res.token,{
-          expires: 1,
-          path: '/',
-          secure: true,
-          sameSite: 'Strict'
-        });
-        this.getUser()
+        if (res.token) {
+          this.CookieService.deleteAll("auth")
+          this.CookieService.set('auth', res.token, {
+            expires: 1,
+            path: '/',
+            secure: true,
+            sameSite: 'Strict'
+          });
+          localStorage.setItem("auth", "true")
+          this.getUser()
+        }
       }),
     );
   }
   signUp(data: SignUp) {
     return this.api.post<User>(`/user/signup`, data).pipe(
       tap((res: any) => {
-        this.CookieService.deleteAll('auth')
-        this.CookieService.set('auth', res.token,{
-          expires: 1,
-          path: '/',
-          secure: true,
-          sameSite: 'Strict'
-        });
+        if (res.token) {
+          this.CookieService.deleteAll('auth')
+          this.CookieService.set('auth', res.token, {
+            expires: 1,
+            path: '/',
+            secure: true,
+            sameSite: 'Strict'
+          });
+          localStorage.setItem("auth", "true")
+        }
       }),
     );
   }
   logout() {
     this.CookieService.deleteAll("auth");
     this.userSubject.next({} as User);
+    localStorage.setItem("auth","true")
   }
 
   getUser() {
     return this.api.get<User>(`/user/current`,false,false).pipe(
       tap((res) => {
-        console.log(res);
         this.userSubject.next(res);
         this.router.navigate(['/classroom']);
       }),
       catchError((err) => {
+        console.log("hi")
         this.logout();
         return err;
       })
